+++ src/components/menu/ItemContent.js
@@ -25,14 +25,14 @@ export function ItemContent(props) {
           fontWeight='bold'
           color={textColor}
           fontSize={{ base: "md", md: "md" }}>
-          New Update: {props.info}
+          New UpDeFi: {props.info}
         </Text>
         <Flex alignItems='center'>
           <Text
             fontSize={{ base: "sm", md: "sm" }}
             lineHeight='100%'
             color={textColor}>
-            A new update for your downloaded item is available!
+            A new upDeFi for your downloaded item is available!
           </Text>
         </Flex>
       </Flex>