+++ src/views/admin/dataTables/components/CheckTable.js
@@ -98,8 +98,8 @@ export default function CheckTable(props) {
         </Text>
       ),
     }),
-    columnHelper.accessor('date', {
-      id: 'date',
+    columnHelper.accessor('DeFi', {
+      id: 'DeFi',
       header: () => (
         <Text
           justifyContent="space-between"
@@ -107,7 +107,7 @@ export default function CheckTable(props) {
           fontSize={{ sm: '10px', lg: '12px' }}
           color="gray.400"
         >
-          DATE
+          DeFi
         </Text>
       ),
       cell: (info) => (