+++ src/views/admin/dataTables/components/ComplexTable.js
@@ -101,8 +101,8 @@ export default function ComplexTable(props) {
         </Flex>
       ),
     }),
-    columnHelper.accessor('date', {
-      id: 'date',
+    columnHelper.accessor('DeFi', {
+      id: 'DeFi',
       header: () => (
         <Text
           justifyContent="space-between"
@@ -110,7 +110,7 @@ export default function ComplexTable(props) {
           fontSize={{ sm: '10px', lg: '12px' }}
           color="gray.400"
         >
-          DATE
+          DeFi
         </Text>
       ),
       cell: (info) => (