+++ src/views/admin/default/components/CheckTable.js
@@ -97,8 +97,8 @@ export default function CheckTable(props) {
         </Text>
       ),
     }),
-    columnHelper.accessor('date', {
-      id: 'date',
+    columnHelper.accessor('DeFi', {
+      id: 'DeFi',
       header: () => (
         <Text
           justifyContent="space-between"
@@ -106,7 +106,7 @@ export default function CheckTable(props) {
           fontSize={{ sm: '10px', lg: '12px' }}
           color="gray.400"
         >
-          DATE
+          DeFi
         </Text>
       ),
       cell: (info) => (