+++ CHANGELOG.md
@@ -4,8 +4,8 @@
 
 - Most vulnerabilities removed, besides those cause by `react-scripts`. We kept this depedency due to the fact that there are
 many users who still use it, and there is already a Next.js version for thos who want to migrate from `react-scripts`.
-- Updated to React 18.
-- Updated react-table to Tanstack V8.
+- UpDeFid to React 18.
+- UpDeFid react-table to Tanstack V8.
 
 ## [1.3.0] 2023-05-06